@@ -5,7 +5,6 @@
   "dependencies": {
     "@emotion/react": "^11.11.4",
     "@emotion/styled": "^11.11.5",
-    "@material-ui/core": "^4.12.4",
     "@mui/icons-material": "^5.10.9",
     "@mui/lab": "^5.0.0-alpha.166",
     "@mui/material": "^5.16.0",
@@ -17,9 +16,9 @@
     "@testing-library/user-event": "^13.5.0",
     "dotenv": "^16.0.3",
     "formik": "^2.2.9",
-    "react": "^18.2.0",
+    "react": "^18.3.1",
     "react-confetti": "^6.1.0",
-    "react-dom": "^18.2.0",
+    "react-dom": "^18.3.1",
     "react-dropzone": "^14.2.3",
     "react-helmet-async": "^2.0.4",
     "react-redux": "^8.0.4",
